GH-20 Refactoring
